pyproject.toml
[project]
name = "analyzer"
version = "0.1.0"
description = "AI-powered document analysis system"
authors = [
    {name = "Seyran Petrosyan",email = "seyranph@gmail.com"}
]
readme = "README.md"
requires-python = "^3.9"
package-mode = false
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "openai (>=1.99.9,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pinecone (>=7.3.0,<8.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.30,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "tiktoken (>=0.11.0,<0.12.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
